backend/requirements.txt
starlette==0.37.2
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
